+++ hobbyfield/src/main/java/com/hobbyfield/app/comm/mapper/ImgFileMapper.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.hobbyfield.app.comm.mapper.ImgFileMapper">
+	<select id="selectImgFile">
+	
+	</select>
+	<insert id="insertImgFile" parameterType="ImgFileVO">
+		Insert 
+	</insert>
+	
+	<update id="updateImgFile">
+	
+	</update>
+	
+	<delete id="deleteImgFile">
+	
+	</delete>
+</mapper>
